ga_core renaming to Green_Algorithms_core and slurm imp. of WM base class
