+++ Green_Algorithms_core/models/job_emissions_model.py
