+++ Green_Algorithms_core/processing/energy.py
